main/modelに構造体を移行
